Fix Float32

+++ fastrand/float.go
@@ -1,5 +1,6 @@
 package fastrand
 
+// Float64 returns a random float64 in [0, 1).
 func (rng *RNG) Float64() float64 {
 retry:
 	f := float64(rng.Int63()) / (1 << 63)
@@ -9,9 +10,10 @@ retry:
 	return f
 }
 
+// Float32 returns a random float32 in [0, 1).
 func (rng *RNG) Float32() float32 {
 retry:
-	f := float32(rng.Float32())
+	f := float32(rng.Float64())
 	if f == 1 {
 		goto retry
 	}
